connecting to db revised version
