+++ .sequelizerc
@@ -1,0 +1,8 @@
+const path = require('path')
+
+module.exports = {
+    config: path.resolve('./src/database/config', 'config.js'),
+    'models-path': path.resolve('./src/database/models'),
+    'seeders-path': path.resolve('./src/database/seeders'),
+    'migrations-path': path.resolve('./src/database/migrations'),
+}
