+++ package.json
@@ -3,9 +3,9 @@
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
-  "type": "module",
   "scripts": {
-    "start": " node ./src/index.js"
+    "build": "babel src --out-dir dist",
+    "start": "npm run build && node dist/index.js"
   },
   "repository": {
     "type": "git",
@@ -18,7 +18,7 @@
   },
   "homepage": "https://github.com/atlp-rwanda/elite-bn-be.git",
   "dependencies": {
-    "@babel/core": "^7.17.12",
+    "@babel/register": "^7.17.7",
     "dotenv": "^16.0.1",
     "express": "^4.18.1",
     "heroku": "^7.60.2",
@@ -27,5 +27,12 @@
     "pg-hstore": "^2.3.4",
     "sequelize": "^6.19.2",
     "sequelize-cli": "^6.4.1"
+  },
+  "devDependencies": {
+    "@babel/cli": "^7.17.10",
+    "@babel/core": "^7.18.0",
+    "@babel/preset-env": "^7.18.0",
+    "babel-plugin-add-module-exports": "^1.0.4",
+    "babel-register": "^6.26.0"
   }
 }
