+++ pull_request_connect_t0_db.md
@@ -1,0 +1,11 @@
+####What does this PR do?
+set all tools necesary for connecting to database
+
+####how should this be manually tested?
+after cloning a repo cd into it and RUN npm run start
+
+####any background context you want to provide?
+can coneect to postgres db and i have provided index for models
+
+####what are the relevant pivotal tracker stories?
+#182201508
