+++ src/database/config/babelHook.js
@@ -1,0 +1,3 @@
+require('babel-core/register');
+
+module.exports = require('./config');
